# src/Homeworks/HW9/openapi.yaml
openapi: 3.0.0
info:
  title: Заказ ресурсов на облаке
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1/
    description: Dev server
paths:
  /clouds:
    get:
      summary: Метод получения списка ресурсов на облако
      tags:
        - Clouds
      operationId: getAllClouds
      responses:
        "200":
            description: Успешный ответ со списком ресурсов в облаке
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Clouds"
        "default":
            description: Все остальное
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Error"
    post:
      summary: Создание заказа на облако
      tags:
        - Clouds
      operationId: CreateCloud
      requestBody:
        content:
          adplication/json:
            schema:
              $ref: "#/components/schemas/Error"
      responses:
        "200":
            description: Успешный ответ со списком ресурсов в облаке
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Clouds"
        "default":
            description: Все остальное
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Error"
  /clouds/{cloud_id}:
    get:
      summary: Метод получения заказа на облако по ID
      tags:
        - Clouds
      operationId: getCloudById
      parameters:
        - name: cloud_id
          in: path
          required: true
          description: Индентификатор заказа облака
          schema:
            type: string
          example: f102b615
      responses:
        "200":
            description: Успешный ответ с заказом облака по ID
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Cloud"
        "404":
            description: Заказ с ID не найден
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Error"
        "default":
            description: Все остальное
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Error"
    delete:
      summary: Метод отмены заказа на облако по ID
      tags:
        - Clouds
      operationId: cancelCloudById
      parameters:
        - name: cloud_id
          in: path
          required: true
          description: Индентификатор заказа облака
          schema:
            type: string
          example: f102b615
      responses:
        "200":
            description: Успешный ответ с заказом облака по ID
            content:
              adplication/json: {}
        "404":
            description: Заказ с ID не найден
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Error"
        "default":
            description: Все остальное
            content:
              adplication/json:
                schema:
                  $ref: "#/components/schemas/Error"
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
  /user/createWithList:
    post:
      tags:
        - user
      summary: Creates list of users with given input array
      description: Creates list of users with given input array
      operationId: createUsersWithListInput
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: successful operation
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: false
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied
  /user/logout:
    get:
      tags:
        - user
      summary: Logs out current logged in user session
      description: ''
      operationId: logoutUser
      parameters: []
      responses:
        default:
          description: successful operation
  /user/{username}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be deleted
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: successful operation
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
components:
  schemas:
    Cloud:
      type: object
      required:
        - id_client
        - OS
        - RAM
        - CPU
        - HDD
      properties:
        cloud_id:
          type: string
          example: f102b615
        id_client:
          type: string
          example: f102b615
        OS:
          type: string
          enum:
            - Windows
            - Linux
          description: Операционная система сервера
        RAM:
          type: string
          example: 128
          description: Объем оперативной памяти
        CPU:
          type: string
          example: 8
          description: Количество ядер процессора
    Clouds:
      type: array
      items:
        $ref: "#/components/schemas/Cloud"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
      xml:
        name: user
    
    
    
    
    
    
    
    
    
    